+++ lib/EPerfCallbacks.cpp
@@ -90,7 +90,7 @@ int cpp_callback_EPerfStopTimer(EPerf *e, int KernelID, int DeviceID) {
 	return E_OK;
 }
 
-int cpp_callback_EPerfAddKernelDataVolumes(EPerf *e, int KernelID, int DeviceID, long long inBytes, long long outBytes) {
+int cpp_callback_EPerfAddKernelDataVolumes(EPerf *e, int KernelID, int DeviceID, int64_t inBytes, int64_t outBytes) {
 	try {
 		e->addKernelDataVolumes(KernelID, DeviceID, inBytes, outBytes);
 	} catch (std::invalid_argument &e) {
